requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
websockets>=11.0
SimConnect>=0.4.26
pydantic>=2.0
httpx>=0.25.0

# Optional: for voice recognition (Whisper)
openai-whisper>=20231117

# Optional: for desktop GUI app (run desktop_app.py)
# pywebview>=5.0
